Improve rendering UI and add clean up functionality
The one wrinkle is that upload requests can be rejected. In principle,
sending CLEANUP in such cases would delete rendered images. But
actually, upload requests are only rejected if the requested image
didn't exist, so this is not an issue, practically speaking.

Also fixed an issue where RenderJobs would write to a file in a
possibly nonexistent directory, causing a FileNotFoundError.